refactor(broker): add PartitionManager
Move construction logic for partition service from AtomixFactory to this class
