+++ broker/src/main/java/io/camunda/zeebe/broker/clustering/atomix/AtomixFactory.java
@@ -17,15 +17,13 @@ import io.atomix.core.Atomix;
 import io.atomix.core.AtomixBuilder;
 import io.atomix.core.AtomixConfig;
 import io.atomix.raft.partition.RaftPartitionGroup;
-import io.atomix.raft.partition.RaftPartitionGroup.Builder;
 import io.atomix.utils.net.Address;
 import io.camunda.zeebe.broker.Loggers;
+import io.camunda.zeebe.broker.partitioning.PartitionManager;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
 import io.camunda.zeebe.broker.system.configuration.ClusterCfg;
 import io.camunda.zeebe.broker.system.configuration.DataCfg;
 import io.camunda.zeebe.broker.system.configuration.MembershipCfg;
-import io.camunda.zeebe.broker.system.configuration.NetworkCfg;
-import io.camunda.zeebe.logstreams.impl.log.ZeebeEntryValidator;
 import io.camunda.zeebe.snapshots.ReceivableSnapshotStoreFactory;
 import io.camunda.zeebe.util.FileUtil;
 import java.io.File;
@@ -90,71 +88,9 @@ public final class AtomixFactory {
     }
 
     final RaftPartitionGroup partitionGroup =
-        createRaftPartitionGroup(configuration, rootDirectory, snapshotStoreFactory);
+        PartitionManager.buildRaftPartitionGroup(configuration, snapshotStoreFactory);
 
     return atomixBuilder.withPartitionGroup(partitionGroup).build();
-  }
-
-  private static RaftPartitionGroup createRaftPartitionGroup(
-      final BrokerCfg configuration,
-      final String rootDirectory,
-      final ReceivableSnapshotStoreFactory snapshotStoreFactory) {
-
-    final File raftDirectory = new File(rootDirectory, AtomixFactory.GROUP_NAME);
-    try {
-      FileUtil.ensureDirectoryExists(raftDirectory.toPath());
-    } catch (final IOException e) {
-      throw new UncheckedIOException("Failed to create raft directory", e);
-    }
-
-    final ClusterCfg clusterCfg = configuration.getCluster();
-    final var experimentalCfg = configuration.getExperimental();
-    final DataCfg dataCfg = configuration.getData();
-    final NetworkCfg networkCfg = configuration.getNetwork();
-
-    final Builder partitionGroupBuilder =
-        RaftPartitionGroup.builder(AtomixFactory.GROUP_NAME)
-            .withNumPartitions(clusterCfg.getPartitionsCount())
-            .withPartitionSize(clusterCfg.getReplicationFactor())
-            .withMembers(getRaftGroupMembers(clusterCfg))
-            .withDataDirectory(raftDirectory)
-            .withSnapshotStoreFactory(snapshotStoreFactory)
-            .withMaxAppendBatchSize((int) experimentalCfg.getMaxAppendBatchSizeInBytes())
-            .withMaxAppendsPerFollower(experimentalCfg.getMaxAppendsPerFollower())
-            .withHeartbeatInterval(clusterCfg.getHeartbeatInterval())
-            .withElectionTimeout(clusterCfg.getElectionTimeout())
-            .withEntryValidator(new ZeebeEntryValidator())
-            .withFlushExplicitly(!experimentalCfg.isDisableExplicitRaftFlush())
-            .withFreeDiskSpace(dataCfg.getFreeDiskSpaceReplicationWatermark())
-            .withJournalIndexDensity(dataCfg.getLogIndexDensity())
-            .withPriorityElection(experimentalCfg.isEnablePriorityElection())
-            .withRequestTimeout(experimentalCfg.getRaft().getRequestTimeout())
-            .withMaxQuorumResponseTimeout(experimentalCfg.getRaft().getMaxQuorumResponseTimeout())
-            .withMinStepDownFailureCount(experimentalCfg.getRaft().getMinStepDownFailureCount());
-
-    final int maxMessageSize = (int) networkCfg.getMaxMessageSizeInBytes();
-
-    final var segmentSize = dataCfg.getLogSegmentSizeInBytes();
-    if (segmentSize < maxMessageSize) {
-      throw new IllegalArgumentException(
-          String.format(
-              "Expected the raft segment size greater than the max message size of %s, but was %s.",
-              maxMessageSize, segmentSize));
-    }
-
-    partitionGroupBuilder.withSegmentSize(segmentSize);
-
-    return partitionGroupBuilder.build();
-  }
-
-  private static List<String> getRaftGroupMembers(final ClusterCfg clusterCfg) {
-    final int clusterSize = clusterCfg.getClusterSize();
-    // node ids are always 0 to clusterSize - 1
-    final List<String> members = new ArrayList<>();
-    for (int i = 0; i < clusterSize; i++) {
-      members.add(Integer.toString(i));
-    }
-    return members;
   }
 
   private static NodeDiscoveryProvider createDiscoveryProvider(
